MESH-234: Add connected device notification for Mesh
Reason for change: This change will add connected device notification to
the Plume NOC

Test Procedure: Verified that the connect/disconnect messages are received and
a message is sent to Plume.

Risks: Low.

Change-Id: I7e9c8f6c748144697a75ec3ac4ef42c84c7ab4c0
Signed-off-by: scott_angel@cable.comcast.com

RDKB-13320: Fix multple connect messages to Webpa

Reason for change: The cloud team has an issue with the STB sending multiple
connect messages for the client devices.

Test Procedure: Verified that only a single connect message is sent to WebPA.
Mesh will receive all messages for syncing to the NOC.

Risks: Low

Change-Id: I54663d5ba2ee541d11430345da4d1e4ee82eadee
Signed-off-by: scott_angel@cable.comcast.com