RDKB-38092 : Notify Component to store notification list for recovery
Reason for change: When notify_comp is not storing notification list in some file, it is in local memory. So, if it crashes and restart , it will not have list of registered parameters, implemented a new solution, to store it in some file and read it from that file whenever notify_comp comes up.

Test Procedure: Kill notify_comp process. Once self heal restarts it, the PA notifications should happen as before for a respective parameter change. The NOTIFYLog.txt.0 can be checked to see if the file has been reloaded.

Risks: None

Signed-off-by: kavya_chowdahallisuresh@comcast.com
(cherry picked from commit 99045567c921cb3dc9fcc4dc94687a1b8e6caedb)
Change-Id: I23effbdd23ecf0d03c67f23be9ee3d28b3cd1158